requirements.txt
flask==2.1.2
flask-bcrypt==1.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.32
werkzeug==2.1.0
requests==2.28.1
numpy==1.24.3
pandas==1.5.3
pymysql==1.0.3
bcrypt~=4.2.1
botocore~=1.35.92
<<<<<<< HEAD
Faker~=33.3.1
python-dotenv==1.0.0
=======
Faker~=33.3.1
>>>>>>> 1b5facb905574e4ccadc1d61226ab5456df3d673
